downloaded the css> style.css,js>script.js and index.html from bootstrapzero.com (free). created a new directory public to serve the static content in it like css/js, I moved the css and js in to public directory.
 Created a new directory srv/views and moved the index.html in it.

Since we have CDNs as href for bootstrap and jquery in the index.html, we want to have bootstrap and jquery locally.
So we installed bower which will take care of these frontend stuff and get it downloaded for us, it works the same way as
npm works, npm is for packages dealing with backend and bower is for packages dealing with frontend like bootstrap etc, so it has
its json file as well like npm has.
we created a file .bowerrc which tells bower where to download/save packages (otherwise by default it will get installed package in to
 bower_components directory), since bootstrap/jquery deals with the frontend/static stuff
we configured .bowerrc with public/lib folder. So public folder has stuff which we added manually and lib would have stuff from bower.

npm install bower -g # globally install
bower install --save --allow-root bootstrap   ## since bootstrap has jquery as dependencies, it will also get installed.