I added the variable sideMenu in the app.js and assigned it values in json format.
Then in app.js where we are calling the module via require(...) I passed this variable i.e. sideMenu.

Now in each file which is using sideMenu I enclose the routes in a function which takes the sideMenu variable as a argument called router and then instead of exporting the express.Router() variable,

I exported the function router which returns the express.Router() variable